pyproject.toml
[tool.poetry]
name = "project-airflow"
version = "0.1.0"
description = ""
authors = ["Aurelio <aurelio.boelter@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
apache-airflow = "^3.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
